requirements.txt
pydantic==2.3.0
pyhumps==3.8.0
requests==2.31.0
